Add client method to query for job results
The `startCommand()` method returns a job id, the results of which
can now be queried using the new `getJobResult()` method.

This query needs to result in a GET request to `/job/<jobid>`, but
the `Connection` class' `getResult(String data)` method would only
produce POST requests. So I added a `getResult()` method which
produces a GET request and adapted the connection implementations
accordingly.