Add support for `pkg' module methods and tests
`pkg' module is a proxy for the underlining
distro of choice where SaltStack is installed.
So not every package manager supports every
operation.

In this case, I have used SaltStack develop (from
git) and aptpkg frontend.

Add support for:
- upgrade_available, which returns whether a
package has an upgrade available from distro
repo
- latest_version, which return the latest version
of the named package available for upgrade.

CAVEAT: if more than one package name is
specified, a dict of name/version
pairs is returned.

This differentiates the implementation when we
call this method with one argument or with
varargs arguments.