Add support for locate module
Adds the full commands for locate module, which includes:
- locate
- updated
- stats
- version

Also this commit adds a test class and an example executable class using "locate"
command.

It was tested with version 2015.8.7 which for me the locate module was not working
and I had to fix the python "file ﻿salt/modules/locate.py":
﻿111,113c111,112
<         if option in kwargs:
<             if bool(kwargs[option]) is True:
<                 options += toggles[option]
---
>         if bool(kwargs[option]) is True:
>             options += toggles[option]

The style of the code generated tries to be coherent with the existent one.