Fix error handling issues with JSON null Collections
This is a first quick try to fix some errors and issues when reading
null Collections.
With the new error handling when a null value is read from a JSON a
GenericSaltError is thrown with a JsonNull.INSTANCE as its content.
But if a Collection has a null value, no error is given and instead a
"result" is returned in the Xor class (the "right" field) with a null
value which is not consistent with the rest of the types and also
makes the application crash when trying to call a "result.get()
.toString()".
This commit adds a new AdapterFactory for an Adapter for Collections
which behaves as the other defined Adapters: if a null value is read
an exception is thrown and catched in the XorAdapter class. This
class is basically a copy of the Gson one with minor changes so
please, review it just in case I did something wrong.