Added local_batch support with tests.
Refactored some code in LocalCall because callSync and callSyncBatch
used nearly the same logic. They now both use a helper function called
callSyncBatchable.

Added a test for callSyncBatch that checks the request body. Also added
a test for callSync that checks the request body to ensure it is not
accidentally using batching when calling callSyncBatchable.